Cleaned up branch. Reusing simpler embedder engine without corpus/search cli references, removed upload-semantic-search scripts for now, to simplify PR, and other misc cleanup
