+++ tasks/Dockerfile
@@ -22,9 +22,6 @@ COPY ./tasks/requirements.txt /tasks/requirements.txt
 WORKDIR /tasks
 RUN pip install -r requirements.txt
 
-# TODO This did not exactly work:
-# RUN python -c "from sentence_transformers import SentenceTransformer; SentenceTransformer('all-mpnet-base-v2')"
-
 COPY ./tasks /tasks
 
 ENV PYTHONPATH "${PYTHONPATH}:/api"
